Remove hub/util/index.py (moved)
